added extra checks in add components

+++ src/commands/add.ts
@@ -1,7 +1,7 @@
 import chalk from "chalk"
 import { Command } from "commander"
 import path from "path"
-import fs from "fs"
+import fs, { mkdirSync } from "fs"
 import ora from "ora"
 import { test } from "../components/test"
 
@@ -27,27 +27,28 @@ const addComponents = async (component: any) => {
         return addComponents.fail("This component doesn't exist!")
     }
     const componentPath = path.resolve(process.cwd(), `src/components/ui/${component}.tsx`)
-    // fs.readFile(componentPath, 'utf8', (err, data) => {
-    //     if (err) {
-    //         if (err.code === 'ENOENT') {
-    //             addComponents.fail()
-    //             console.error('File not found:', componentPath);
-    //         } else {
-    //             addComponents.fail()
-    //             console.error('Error reading file:', err);
-    //         }
-    //         return;
-    //     }
-    // });
-    try {
-        console.log(componentPath)
+    const componentsDirExist = path.resolve(process.cwd(), `src/components`)
+    const UIDirExist = path.resolve(process.cwd(), `src/components/ui`)
+    if (!fs.existsSync(componentsDirExist)) {
+        mkdirSync(componentsDirExist)
+        mkdirSync(UIDirExist)
         fs.writeFile(componentPath, componentFile.code, () => {
             addComponents.succeed()
         })
-    } catch (error) {
-        console.log(error)
-        addComponents.fail("This component doesn't exist!")
+    } else if (!fs.existsSync(UIDirExist)) {
+        mkdirSync(UIDirExist)
+        fs.writeFile(componentPath, componentFile.code, () => {
+            addComponents.succeed()
+        })
+    } else {
+        try {
+            fs.writeFile(componentPath, componentFile.code, () => {
+                addComponents.succeed()
+            })
+        } catch (error) {
+            console.log(error)
+            addComponents.fail("This component doesn't exist!")
+        }
     }
-
 }
